pyproject.toml
[tool.poetry]
name = "academia-jedi"
version = "0.59.0"
description = ""
authors = ["Britodfbr <britodfbr@gmail.com>"]
packages = [
  {include='incolume', from=''}
]
[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.3"
SQLAlchemy = "^1.4.39"
pandas = "^1.4.3"
requests = "^2.28.1"
python-dotenv = "^0.20.0"
bs4 = "^0.0.1"
lxml = "^4.9.1"
jupyter = "^1.0.0"
getch = "^1.0"
flask-restful = "^0.3.9"
pony = "^0.7.16"
pydantic = "^1.10.2"
unidecode = "^1.3.6"
pysimplegui = "^4.60.4"
pysimpleguiqt = "^0.35.0"
easygui = "^0.98.3"
openpyxl = "^3.0.10"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = {extras = ["jupyter"], version = "^22.12.0"}
Faker = "^13.15.1"
dynaconf = "^3.1.11"
fake = "^0.8"

[tool.poetry.group.dev.dependencies]
markdown = "^3.4.1"
selectolax = "^0.3.11"
rstr = "^3.2.0"
aiohttp = "^3.8.3"
pwinput = "^1.0.2"
fastparquet = "^2022.12.0"
pyarrow = "^10.0.1"
pypdf2 = "^3.0.1"
pymupdf = "^1.21.1"
pdfplumber = "^0.7.6"
pdfminer3 = "^2018.12.3.0"
pytesseract = "^0.3.10"
pdf2image = "^1.16.2"
pillow = "^9.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
